refactor: move edit embed modal to a separate component
